move tracklog query logic to dedicated service
