--- grails-app/services/io/xh/hoist/track/TrackLogAdminService.java
@@ -1,4 +1,0 @@
-package io.xh.hoist.track;
-
-public class TrackLogAdminService {
-}
